Add another layer of abstraction (MsalTokenProvider)
