+++ pom.xml
@@ -39,7 +39,7 @@
         <!-- Compile dependencies -->
         <slf4j.version>1.8.0-beta4</slf4j.version>
         <commons-lang3.version>3.9</commons-lang3.version>
-        <msal4j.version>1.7.0</msal4j.version>
+        <msal4j.version>1.8.1</msal4j.version>
         <httpclient.version>4.5.8</httpclient.version>
         <httpcore.version>4.4.11</httpcore.version>
         <json.version>20190722</json.version>
